Imported only required d3 libraries. Refactored some of the code.
ci: version bump to 0.77.6

Refactored styles.

Added some typing and styling.

ci: version bump to 0.77.7

ci: version bump to 0.77.8

Removed renderObject from Step function renderer.

ci: version bump to 0.77.9
