+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "@toyota-research-institute/lakefront",
   "description": "React UI Components",
-  "version": "0.77.5",
+  "version": "0.77.9",
   "main": "dist/index.cjs.js",
   "module": "dist/index.esm.js",
   "homepage": "https://github.com/ToyotaResearchInstitute/lakefront",
@@ -28,7 +28,9 @@
     "@emotion/react": "^11.1.5",
     "@emotion/styled": "^11.1.5",
     "copyfiles": "^2.4.1",
-    "d3": "^7.0.1",
+    "d3-selection": "^3.0.0",
+    "d3-shape": "^3.0.1",
+    "d3-zoom": "^3.0.0",
     "dagre-d3": "^0.6.4",
     "query-string": "^7.0.0",
     "ramda": "^0.27.1",
@@ -53,6 +55,10 @@
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.5",
     "@testing-library/react-hooks": "^7.0.0",
+    "@types/d3-selection": "^3.0.1",
+    "@types/d3-shape": "^3.0.2",
+    "@types/d3-zoom": "^3.0.1",
+    "@types/dagre-d3": "^0.6.3",
     "@types/jest": "^26.0.20",
     "@types/ramda": "github:types/npm-ramda#dist",
     "@types/react": "^17.0.3",
