+++ src/StepFunctionRenderer/util/index.ts
@@ -1,0 +1,2 @@
+export * from './graphUtil';
+export * from './stepFunctionUtil';
